Move into ObjectProcessor
