--- src/test/java/edu/si/fcrepo/AkubraIDConverterTest.java
@@ -1,92 +1,0 @@
-/*
- * Copyright 2017 Smithsonian Institution.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License.You may obtain a copy of
- * the License at: http://www.apache.org/licenses/
- *
- * This software and accompanying documentation is supplied without
- * warranty of any kind. The copyright holder and the Smithsonian Institution:
- * (1) expressly disclaim any warranties, express or implied, including but not
- * limited to any implied warranties of merchantability, fitness for a
- * particular purpose, title or non-infringement; (2) do not assume any legal
- * liability or responsibility for the accuracy, completeness, or usefulness of
- * the software; (3) do not represent that use of the software would not
- * infringe privately owned rights; (4) do not warrant that the software
- * is error-free or will be maintained, supported, updated or enhanced;
- * (5) will not be liable for any indirect, incidental, consequential special
- * or punitive damages of any kind or nature, including but not limited to lost
- * profits or loss of data, on any basis arising from contract, tort or
- * otherwise, even if any of the parties has been warned of the possibility of
- * such loss or damage.
- *
- * This distribution includes several third-party libraries, each with their own
- * license terms. For a complete copy of all copyright and license terms, including
- * those of third-party libraries, please see the product release notes.
- */
-
-package edu.si.fcrepo;
-
-import static edu.si.fcrepo.AkubraIDConverter.getBlobId;
-import static org.junit.Assert.assertEquals;
-
-import java.net.URI;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.junit.MockitoJUnitRunner;
-
-import static edu.si.fcrepo.AkubraIDConverter.FEDORA_URI;
-
-/**
- * AkubraIDConverter unit tests
- *
- * @author whikloj
- */
-@RunWith(MockitoJUnitRunner.class)
-public class AkubraIDConverterTest {
-
-    @Test
-    public void testObjectIDNoPrefix() {
-        final String objectID = "testid:123";
-        final URI expected = URI.create(FEDORA_URI + objectID);
-        final URI output1 = getBlobId(objectID);
-        assertEquals("Incorrect object blobId built from string.", expected, output1);
-
-        final URI output2 = getBlobId(URI.create(objectID));
-        assertEquals("Incorrect object blobId built from URI.", expected, output2);
-    }
-
-    @Test
-    public void testObjectID() {
-        final String objectID = FEDORA_URI + "testid:123";
-        final URI expected = URI.create(objectID);
-        final URI output1 = getBlobId(objectID);
-        assertEquals("Incorrect object blobId built from string.", expected, output1);
-
-        final URI output2 = getBlobId(URI.create(objectID));
-        assertEquals("Incorrect object blobId built from URI.", expected, output2);
-    }
-
-    @Test
-    public void testDsIDNoPrefix() {
-        final String objectID = "testid:123+DC+DC.0";
-        final URI expected = URI.create(FEDORA_URI + "testid:123/DC/DC.0");
-        final URI output1 = getBlobId(objectID);
-        assertEquals("Incorrect datastream blobId built from string.", expected, output1);
-
-        final URI output2 = getBlobId(URI.create(objectID));
-        assertEquals("Incorrect datastream blobId built from URI.", expected, output2);
-    }
-
-    @Test
-    public void testDsID() {
-        final String objectID = FEDORA_URI + "testid:123+DC+DC.0";
-        final URI expected = URI.create(FEDORA_URI + "testid:123/DC/DC.0");
-        final URI output1 = getBlobId(objectID);
-        assertEquals("Incorrect datastream blobId built from string.", expected, output1);
-
-        final URI output2 = getBlobId(URI.create(objectID));
-        assertEquals("Incorrect datastream blobId built from URI.", expected, output2);
-    }
-}
